+++ fmbuf/fmbuf.h
@@ -1,3 +1,7 @@
+/**
+ *  Light Weight Memory Buffer, which can be used for Ring-Buffer or A Simple Array
+ */
+
 #ifndef MBUF_H_FINAL
 #define MBUF_H_FINAL
 
@@ -11,28 +15,29 @@ fmbuf*  fmbuf_create(size_t size);
 void    fmbuf_delete(fmbuf* pbuf);
 
 // public interface 
-int     fmbuf_size(fmbuf* pbuf);
-int     fmbuf_used(fmbuf* pbuf);
+size_t  fmbuf_size(fmbuf* pbuf);
+size_t  fmbuf_used(fmbuf* pbuf);
 void    fmbuf_clear(fmbuf* pbuf);
 void*   fmbuf_get_head(fmbuf* pbuf);
 void*   fmbuf_get_tail(fmbuf* pbuf);
 
-// ----------------As a Loop Buff------------------------
+// ----------------As a Ring Buff------------------------
 //push data into mbuf : if tail free space not enough that continue write from start
 //return 0/1 : 0-> sucess 1->fail
 int     fmbuf_push(fmbuf* pbuf, const void* data, size_t size);
 int     fmbuf_pop(fmbuf* pbuf, void* data, size_t size);
+
 void*   fmbuf_vpop(fmbuf* pbuf, void* data, size_t size);
 void*   fmbuf_getraw(fmbuf* pbuf, void* data, size_t size);
 void    fmbuf_head_move(fmbuf* pbuf, size_t size);
-int     fmbuf_free(fmbuf* pbuf);
+size_t  fmbuf_free(fmbuf* pbuf);
 
 //-----------------As a Array-----------------------------
 //alloc a buf from mbuf
 //return NULL : alloc failed
-int     fmbuf_head_free(fmbuf* pbuf);
-int     fmbuf_tail_free(fmbuf* pbuf);
-int     fmbuf_total_free(fmbuf* pbuf);
+size_t  fmbuf_head_free(fmbuf* pbuf);
+size_t  fmbuf_tail_free(fmbuf* pbuf);
+size_t  fmbuf_total_free(fmbuf* pbuf);
 
 void*   fmbuf_alloc(fmbuf* pbuf, size_t size);
 void    fmbuf_head_seek(fmbuf* pbuf, int offset);
